chore: updates

+++ vscode/link.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+[ "$(uname -s)" != "Darwin" ] && exit 0
+
+
 rm $HOME/Library/Application\ Support/Code/User/settings.json
 ln -s $HOME/.dotfiles/vscode/settings.json $HOME/Library/Application\ Support/Code/User/settings.json
 rm $HOME/Library/Application\ Support/Code/User/keybindings.json
